iteration1_py/requirements.txt
# Temel bağımlılıklar
pytest>=7.0.0
mypy>=1.0.0

# Opsiyonel (projenin ihtiyacına göre)
pandas>=2.0.0
pyyaml>=6.0